### 2018-08-07 #880 @pjotrp
This is potentially an interesting contribution - especially since it uses Uniprot. The git repo is pretty large, it would be worth throwing out some large files from history (not so light-weight as it needs to be I mean). The species.json file is massive - would it not be better to generate that too?

Question: do you have the website running somewhere so I can try? It does not say in the README.
