### 2018-08-07 b9c3760
Fix list formatting in paper.
