### 2018-08-07 594031d
one more try at fixing list/figure caption
