### 2018-08-08 #880 @serine
@j-andrews7 @pjotrp This looks pretty neat ! worked out of the box in first attempt. not sure of the usage/target audience yet, but certainly great aggregate of things and as an exploratory tools will be great resource.

Having this as a stand along electron app is initial wow factor (at least for me), but not sure sure if this is much of advantage over website/webapp, seems like webapp would have been even more accessible.

Just a comment; paper pdf feels long, while all information is very relevant and useful, perhaps fair chunk can be moved into general docs. Like I said I don't actually know what JOSS policies are. I thought paper.md was meant to abstract like. @arfon is this so?

In summary I'm rather interested in reviewing this paper further.

Cheers,
