### 2018-08-08 #880 @serine
I should probably add that those comments were rather general and don't directly apply to the review process. From what I understand the review process more or less whether the tools installs and works as described. And it did install and worked. I just need to play with it to appreciate the usability and applicability of the tool.

One thing that stood out from reading the paper and the github; "gene annotation" seems to be used a few times. To me gene annotation is annotation of a new - unknown region of the genome with a specific name/feature e.g protein_coding gene. In my head gene annotation usually requires two things 1. the data e.g RNAseq and 2. the genome browser of some sort to browser through the genome and looking for new pile ups of reads where there is no known feature, thereby annotating that region as new_feature_A.

Cheers
