### 2018-08-08 #880 @arfon
> Just a comment; paper pdf feels long, while all information is very relevant and useful, perhaps fair chunk can be moved into general docs. Like I said I don't actually know what JOSS policies are. I thought paper.md was meant to abstract like. @arfon is this so?

This paper is certainly longer than most but I think it's OK. The 1-page figure makes it seem quite long (and might be more appropriate in the docs).
