### 2018-08-08 #880 @j-andrews7
@serine Thanks for the comments. I'm glad it's working well for you. As for length, the guidelines said 250-1000, I think the paper as it currently stands is ~700. I could probably cut it down a bit, I just wanted to establish the need for it well. I'm not sure if I can really resize the figure in Rmarkdown given that I want to keep the high resolution, but I can certainly move it to the docs.

The reason I didn't build it as a web app was so I could hotkey it so that it'd be quicker to run through a column of gene names in excel. I was also trying to design it so it could be used on a quarter/half-screen side by side a table or another browser for those of us stuck with single monitor setups. Not having to host the website is also a plus, and I wanted to take a crack at what Electron could do.

As for the "gene annotation" phrase, I kind of agree, but I saw it used in a similar context several other places (even from places like NCBI and Uniprot).
