### 2018-08-08 #885 @whedon
[REVIEW]: Genotify: Fast, lightweight gene lookup and summarization
**Submitting author:** @j-andrews7 (<a href="http://orcid.org/0000-0002-0780-6248">Jared Andrews</a>)
**Repository:** <a href="https://github.com/j-andrews7/Genotify" target ="_blank">https://github.com/j-andrews7/Genotify</a>
**Version:** v1.2.0
