### 2018-08-08 #885 @j-andrews7
@pjotrp I can't tick them, but they are all unticked currently. The PDF looks fine to me.
