### 2018-08-08 #885 @j-andrews7
@pjotrp Oh, sorry, I misunderstood. Yes, all of the checkboxes are fine. Given that it is a GUI-based application, I don't include automated tests, but it is quite easy to test manually.
