### 2018-08-08 #885 @serine
@j-andrews7 I'll try to play with genotify over the next few days. I'll have more time next week.

In terms of the documentation. Do you think you can write somethings up? Like perhaps "typical" usage or an example of how you've used in the past? I understand that this is an aggregate of diff tools plus some extra bits and is for explanatory purpose, but you obviously have a particular use case, if you can share that "workflow". I think that would be useful, as well as strengthen the argument for the tool.

These are the actual dot points that I need to tick off.

- **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
- **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
- **Automated tests:** Are there automated tests or manual steps described so that the function of the software can be verified?
- **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Also note that **A statement of need** comes up twice on my checklist. under docs and under software paper. I ticked off the one under software paper, but haven't yet ticked off the one under the docs. I think this a minor thing, but all kind of rotates around "functional docs" idea. So if you don't mind putting it all in some sort of docs, that would be very beneficial for the end user. After all paper will remain "static" but docs may change with time.

- **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?

You probably have some idea on how to do docs. But if I can suggest to keep them in the repo i.e under `docs/` in plain text e.g markdown, that I think works best. You can then use something like [mkdocs](https://www.mkdocs.org/) or similar to compile into html and push to [gh-pages](https://pages.github.com/). The advantage of keep them in the repo would be, "everything in one place", but also rather easy to contribute to the docs - clone the repo, write some markdown and PR back. Something that community might want to do as tools growth in popularity.

Cheers
