### 2018-08-09 #885 @j-andrews7
I can definitely write some additional docs, but I am unsure how necessary it would be to create a site for them. Would adding a functional example or two to the README work?
