### 2018-08-09 224d5ba
provide a more illustrative example for the README
