### 2018-08-15 #885 @serine
@j-andrews7 @pjotrp and @arfon All is good and here is my final comment about the tools below. I do have a couple of question regarding finalising review process, will post them as a separate comment next.

Overall I think this is great tool. It'll most likely become a go to app for me when I need to check the gene and associate info. The tool aggregates many different API's nicely making it very easy to get a thorough overview about given gene, including functions, expression and diseases associations. The tools works as described with very quick and simply installation. I've tested it out on my linux (debian distro) and macOS, both worked out of the box with less then a minute of install time required.

Going forward I would like to see some small improvements in the docs, mainly to attract new comers and make super easy to understand what each section is about. I found it a little distracting having to jump between _paper.md_ and _README.md_/docs. I would prefer docs with some table of content and all in one place.
I think adding additional "references links" to each sections websites docs pages for more in depth information would be rather useful. While one can use google for that, it feels like that would really complement the tool on the ease of use aspect.

I'd like to also suggest if small improvements can be made in **Expression** and **Diseases** sections going forward. The second panel in the expression sections - digram view seems to update with every new query, but the table appears to be static with the same content irrespective of the query. I've noticed that if I try to click on few different experiments some show expression and logFC but other show _no data is avaliable_. Can it be that those data sets without expression data simply filtered out from that table?
The Diseases section doesn't redirect to CTDbase, don't know if it should, seems like that would be useful.
Unless I've missed something

Lastly I would also suggest considering webapp on top of existing electron app, mainly to attract more users and give them options. I feel that one would be more likely to test a tool with zero commitments i.e just hit "this" webpage. Regarding website hosting, I'm fairly confident you can do that through gh-page for free and all will be kept in the single repo.

All the best and thanks for the tool !
