### 2018-08-15 #885 @serine
@pjotrp and @arfon

I'm little unsure about

``` Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?```

section. while every code is gotta be tested I'm a little uncertain which tests to suggest for this particular tool? As I've mentioned everything appears to work fine.

Other wise I'm happy to mark this as done

Thanks
