### 2018-08-15 0709a1d
add ToC, update readme
