### 2018-08-15 #885 @j-andrews7
Thanks for the comments @serine. I will continue to update the docs as I make improvements to the tool. As it stands, I'd assume most people might take a quick read over the paper, then mostly just go through the `README`.

As you've noted, some expression experiments don't have data that you might expect - typically these are differential expression experiments that will only display information if the gene is differentially expressed between whatever the two conditions are. Most of the *RNA-Seq Baseline*-type experiments are usually more reliable. Unfortunately, there's not much I can do about it - that's just how the data from **Expression Atlas** is, even through their website. I very much agree that it is worth mentioning in the docs, as is adding a ToC. The experiments table will only change if the species changes - the experiments available for Human remain static, but if you click on a Mouse hit, it should change to display the mouse experiments. If it is not, then you may be hitting a bug, which I'd suggest reporting as an issue in the Genotify repo.

I have updated the `README` to explicitly state those nuances, since I realize they aren't obvious, as well as to add a ToC to better organize it.

I will change the **Disease** section to add a column that directs back to CTDbase as well.

As for the webpage, it's an idea I'll consider, but it will take some effort to de-couple the code from some of the Electron intricacies. `gh-pages` also only lets you publish static HTML files, so you have no control over the backend, i.e. I couldn't install `node.js` or any of the other `js` libraries I'd need for it to work. I would have to host elsewhere, which isn't a big deal, but not quite as convenient. And as you said, I'm hoping the quick install will encourage people to at least give it a try.

Thanks for reviewing, and I'm glad you've found it a useful tool. I will implement the change above (and I've also found a few edge-case bugs since submitting), and prepare to make a release that will be used for the archive.

@pjotrp is there anything I need to do to update the release version for the submission? It's listed as 1.2.0 here, but my archive will be version 1.2.1.
