### 2018-08-15 a89d9ad
add link to ctd base table, fix expression table bug
also updated the changelog
