### 2018-08-15 #885 @serine
@j-andrews7 that's great, thanks heaps for addressing comments.

I think I missed that genotify has a backend, apologies, in that case yes gh-pages is no good. And like I said this is completely up to you, just an idea to think about. The electron app is very good.

Cheers
